requirements.txt
setuptools~=51.1.1
schedule~=0.6.0
GithubArchiver~=0.0.1
PyGithub~=1.54.1